+++ src/platform/Windows/WindowsWindow.cpp
@@ -146,7 +146,7 @@ namespace Fractal
 			{
 				WindowData& data = *static_cast<WindowData*>(glfwGetWindowUserPointer(window));
 
-				MouseScrolledEvent event(offsetX, offsetY);
+				MouseScrolledEvent event(static_cast<float>(offsetX), static_cast<float>(offsetY));
 				data.fEventCallback(event);
 			});
 
@@ -154,7 +154,7 @@ namespace Fractal
 			{
 				WindowData& data = *static_cast<WindowData*>(glfwGetWindowUserPointer(window));
 
-				MouseMovedEvent event(positionX, positionY);
+				MouseMovedEvent event(static_cast<float>(positionX), static_cast<float>(positionY));
 				data.fEventCallback(event);
 			});
 	}
